# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users and Roles",
        "description": "This is a Users and Roles API using Lumen and Swagger",
        "contact": {
            "email": "rhiannon.eldridge.au@gmail.com"
        },
        "license": {
            "name": "BY-NC-SA",
            "url": "https://creativecommons.org/licenses/by-nc-sa/4.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost Server"
        }
    ],
    "paths": {
...
        "/api/users/{user_id}/roles/{role_id}": {
            "get": {
                "tags": [
                    "userRoles"
                ],
                "summary": "Get Role from User",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of all Roles assigned to this User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "userRoles"
                ],
                "summary": "Assign User Role",
                "operationId": "assignUserRole",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of all Roles assigned to this User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "userRoles"
                ],
                "summary": "Delete User Role",
                "operationId": "deleteUserRole",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of all Roles assigned to this User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get All Users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Array of all Users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Create User object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Created"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The specified User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Update User",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update User object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete User",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The User has been deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateUpdateUser": {
                "title": "Create/Update User model",
                "description": "Create/Update User model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of User",
                        "format": "string"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email Address of User",
                        "format": "email"
                    },
                    "password": {
                        "title": "Password",
                        "description": "User password",
                        "format": "string"
                    },
                    "api_token": {
                        "title": "API Token",
                        "description": "API Token for this user",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "Role": {
                "title": "Role",
                "description": "Role schema",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of Role",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of Role",
                        "format": "string"
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Date Role was created",
                        "format": "string"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Date Role was last updated",
                        "format": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User schema",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of User",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of User",
                        "format": "string"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email Address of User",
                        "format": "email"
                    },
                    "api_token": {
                        "title": "API Token",
                        "description": "API Token for this User",
                        "format": "string"
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Date User was created",
                        "format": "string"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Date User was last updated",
                        "format": "string"
                    },
                    "roles": {
                        "title": "Roles",
                        "description": "Roles assigned to this User",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Your API Key",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ],
...
}